requirements.txt
streamlit
numpy
faiss-cpu
transformers
torch
python-dotenv